breaks global nav into separate component files, adds tests
