+++ src/assets/stylesheets/styles.scss
@@ -113,7 +113,6 @@
   background: color('blue-60') linear-gradient(0deg, rgba(0, 0, 0, 0) 25%, rgba(0, 0, 0, 0.2) 100%);
 
   a {
-    color: white; // sets login / logout link color
     gap: 0.5rem;
 
     &.nav-product {
