+++ src/components/auth/LogoutButton.js
@@ -1,3 +1,7 @@
 export function LogoutButton() {
-  return <a href="/logout">Log Out</a>;
+  return (
+    <a href="/logout" className="text-white">
+      Log Out
+    </a>
+  );
 }
